tf binding: squelch msvc warnings

+++ binding.cc
@@ -736,7 +736,7 @@ TF_Tensor* CreateSmallTensor(napi_env env,
     switch (dtype) {
       case TF_FLOAT: {
         float* data = new float[1];
-        data[0] = GetDoubleValue(env, data_js);
+        data[0] = static_cast<float>(GetDoubleValue(env, data_js));
         return TF_NewTensor(dtype, NULL, 0, data, 4, ReleaseSmallHandle, NULL);
       }
 
@@ -763,7 +763,7 @@ TF_Tensor* CreateSmallTensor(napi_env env,
       auto data = new float[data_length];
       for (uint32_t i = 0; i < data_length; ++i) {
         napi_value val = GetElement(env, data_js, i);
-        data[i] = GetDoubleValue(env, val);
+        data[i] = static_cast<float>(GetDoubleValue(env, val));
       }
       return TF_NewTensor(
           dtype, shape, 1, data, 4 * data_length, ReleaseSmallHandle, NULL);
